Add files via upload
To get started with the data collection, we can use the “Jupyter Notebook”  to run some Python code that will help to collect some data. Analyse data
Once We have our dataset, we need to prepare it. The data will be very messy and contain purely text. we will need to perform data cleaning in order to prepare the data for analysis.