Add files via upload
When the data is clean, we performed our own analysis to uncover some insights. As a starting point, we could look at topic modelling, sentiment analysis or wordclouds to provide some insight into the content of the reviews. It is recommended to complete this task using Python, however, we can use any tool that we wish. we can use some of the documentation websites provided in the Resources section below to analyse the data.